// src/frontend/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type BehandlungsartDto = {
    id: string;
    name: string;
    preis: number;
};

export type EingelesenerPatientDto = {
    titel?: string;
    vorname: string;
    nachname: string;
    strasse: string;
    hausnummer: string;
    postleitzahl: string;
    stadt: string;
    geburtstag: string;
};

export type EingelesenesRezeptDto = {
    ausgestelltAm: string;
    rezeptpositionen: Array<EingelesenesRezeptPosDto>;
};

export type EingelesenesRezeptPosDto = {
    anzahl: number;
    behandlungsart: BehandlungsartDto;
};

export type PatientDto = {
    id: string;
    titel?: string;
    vorname: string;
    nachname: string;
    strasse?: string;
    hausnummer?: string;
    plz?: string;
    stadt?: string;
    telMobil?: string;
    telFestnetz?: string;
    email?: string;
    geburtstag?: string;
};

export type PatientForm = {
    titel?: string;
    vorname?: string;
    nachname?: string;
    strasse?: string;
    hausnummer?: string;
    plz?: string;
    stadt?: string;
    telMobil?: string;
    telFestnetz?: string;
    email?: string;
    geburtstag?: string;
};

export type RezeptArztDto = {
    id: string;
    name: string;
};

export type RezeptDto = {
    id: string;
    patient: RezeptPatientDto;
    ausgestelltAm: string;
    ausgestelltVon?: RezeptArztDto;
    preisGesamt: number;
    rechnung?: RezeptRechnungDto;
    positionen: Array<RezeptPosDto>;
};

export type RezeptEinlesenResponse = {
    existingPatient?: PatientDto;
    patient: EingelesenerPatientDto;
    rezept: EingelesenesRezeptDto;
    path: string;
};

export type RezeptPatientDto = {
    id: string;
    vorname: string;
    nachname: string;
};

export type RezeptPosDto = {
    beahandlungsart: BehandlungsartDto;
    anzahl: number;
};

export type RezeptRechnungDto = {
    rechnungsnummer: string;
    status: 'ERSTELLT' | 'OFFEN' | 'BEZAHLT';
};

export type status = 'ERSTELLT' | 'OFFEN' | 'BEZAHLT';

export type CreateFromImageData = {
    body?: {
        file: (Blob | File);
    };
};

export type CreateFromImageResponse = (RezeptEinlesenResponse);

export type CreateFromImageError = unknown;

export type GetPatientenResponse = (Array<PatientDto>);

export type GetPatientenError = unknown;

export type CreatePatientData = {
    body: PatientForm;
};

export type CreatePatientResponse = (PatientDto);

export type CreatePatientError = unknown;

export type DeletePatientData = {
    path: {
        id: string;
    };
};

export type DeletePatientResponse = (unknown);

export type DeletePatientError = unknown;

export type UpdatePatientData = {
    body: PatientForm;
    path: {
        id: string;
    };
};

export type UpdatePatientResponse = (PatientDto);

export type UpdatePatientError = unknown;

export type GetRezepteResponse = (Array<RezeptDto>);

export type GetRezepteError = unknown;

export type ServeFileData = {
    path: {
        filename: string;
    };
};

export type ServeFileResponse = ((Blob | File));

export type ServeFileError = unknown;

export type DeleteRezeptData = {
    path: {
        id: string;
    };
};

export type DeleteRezeptResponse = (unknown);

export type DeleteRezeptError = unknown;